Setup benchmarks to disable hard-link counting where possible.
According to PDUs limitations, it will count each hard-link to a file.
As the correct handling of hard-links isn't free, I think it makes sense
to count hard-links in all tools that support it.

All of those which support it have the `-l` flag added to them. It was chosen
over its long form to avoid making the command-line for `gdu` even longer and reduce
the bar size the least amount possible.

Please note that `ncdu` doesn't allow to disable hard-link counting, while `dust` doesn't
support it either.